backend/requirements.txt
Flask
Flask-PyMongo
firebase-admin
python-dotenv
flask-cors
twilio
requests
Werkzeug==2.2.2